native: Use size_t instead of int where appropriate

+++ src/cc_item.h
@@ -3,11 +3,11 @@
 
 #define ITEM_ID_SIZE 64
 
-struct cc_item {         
-  char  item_id[ITEM_ID_SIZE];
-  int   coconcurrency_count;
-  int   total_count;
-  float similarity;  
-};         
+struct cc_item {
+  char   item_id[ITEM_ID_SIZE];
+  size_t coconcurrency_count;
+  size_t total_count;
+  float  similarity;
+};
 
 #endif
