+++ src/cosine.c
@@ -2,6 +2,6 @@
 
 #include "cc_item.h"
 
-void calculate_cosine(char const *restrict item_id, int itemCount, struct cc_item *restrict cc_items, int cc_items_size){
+void calculate_cosine(char const *restrict item_id, size_t itemCount, struct cc_item *restrict cc_items, size_t cc_items_size){
   /* here be dragons */
 }
