+++ src/jaccard.h
@@ -3,6 +3,6 @@
 
 #include "cc_item.h"
 
-void calculate_jaccard(char const *restrict item_id, int itemCount, struct cc_item *restrict cc_items, int cc_items_size);
+void calculate_jaccard(char const *restrict item_id, size_t itemCount, struct cc_item *restrict cc_items, size_t cc_items_size);
 
 #endif
